Python is the worst language. Ever. Really, it sucks ass. Useless. I want things done today, not tomorrow.
Also, I improved fft demo a bit as ossaudiodev apparently create a buffer I can't change the size of (so all them music is delayed with about half a sec).
Since python sucks monkey balls the change is insignificant, because the execution of the code is way too slow and /dev/dsp can't be read fast enough.

And i made martha actually check that the brightness value isn't out of bounds which it should.